requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-oauth-toolkit==1.3.2
django-oauth2-provider==0.2.6.1
djangorestframework==3.11.0
drf-yasg==1.17.1
fdm==0.2.0
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
plum-dispatch==0.2.0
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
